correccion calendario mis-turnos
